+++ src/frontend/app/app.common.module.ts
@@ -131,8 +131,6 @@ export function getAngulartics2Providers(): any {
         StarDocActionTagService,
         StarDocPlaylistService,
         {provide: PrintService, useClass: SimplePrintService},
-        // TODO if you want pdf replace PrintDialogPdfGenerator by JsPdfGenerator and move jspdf in package.json from optional to dep
-        {provide: PdfGenerator, useClass: PrintDialogPdfGenerator},
         {provide: PdfPrintService, useClass: SimplePdfPrintService}
     ],
     bootstrap: [AppComponent]