+++ src/frontend/app/app.module.ts
@@ -58,6 +58,8 @@ import {
 } from '@dps/mycms-frontend-commons/dist/angular-commons/htmlrenderer/html-toggler.renderer';
 import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {AppService} from './services/app.service';
+import {PdfGenerator} from '@dps/mycms-frontend-commons/dist/angular-commons/services/pdf-print.service';
+import {PrintDialogPdfGenerator} from '@dps/mycms-frontend-commons/dist/angular-commons/services/print-dialog-pdf.generator';
 
 registerLocaleData(localeDe);
 
@@ -83,6 +85,8 @@ registerLocaleData(localeDe);
         PDocNameSuggesterService,
         PDocDescSuggesterService,
         PDocPageDescSuggesterService,
+        // TODO if you want pdf replace PrintDialogPdfGenerator by JsPdfGenerator and move jspdf in package.json from optional to dep
+        {provide: PdfGenerator, useClass: PrintDialogPdfGenerator},
         {provide: AngularMarkdownService, useClass: SpecificAngularMarkdownService},
         {provide: AngularHtmlService, useClass: SpecificAngularHtmlService},
         HtmlLocalLinkRenderer,