+++ src/frontend/app/app.readonly.module.ts
@@ -19,6 +19,9 @@ import {
 } from '@dps/mycms-frontend-commons/dist/angular-commons/htmlrenderer/html-toggler.renderer';
 import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {AppService} from './services/app.service';
+import {MarkdownPadModule} from './mdpad/mdpad.module';
+import {PdfGenerator} from '@dps/mycms-frontend-commons/dist/angular-commons/services/pdf-print.service';
+import {PrintDialogPdfGenerator} from '@dps/mycms-frontend-commons/dist/angular-commons/services/print-dialog-pdf.generator';
 
 registerLocaleData(localeDe);
 
@@ -27,10 +30,13 @@ registerLocaleData(localeDe);
     ],
     imports: [
         AppCommonModule,
+        MarkdownPadModule,
         AppRoutingModule
     ],
     providers: [
         {provide: GenericAppService, useClass: AppService},
+        // TODO if you want pdf replace PrintDialogPdfGenerator by JsPdfGenerator and move jspdf in package.json from optional to dep
+        {provide: PdfGenerator, useClass: PrintDialogPdfGenerator},
         {provide: AngularMarkdownService, useClass: SpecificAngularMarkdownService},
         {provide: AngularHtmlService, useClass: SpecificAngularHtmlService},
         HtmlLocalLinkRenderer,